Avoid a warning

+++ content/number-theory/FastEratosthenes.h
@@ -20,7 +20,7 @@
 const int LIM = 1e6;
 bitset<LIM> isPrime;
 vi eratosthenes() {
-	const int S = round(sqrt(LIM)), R = LIM / 2;
+	const int S = (int)round(sqrt(LIM)), R = LIM / 2;
 	vi pr = {2}, sieve(S+1); pr.reserve(int(LIM/log(LIM)*1.1));
 	vector<pii> cp;
 	for (int i = 3; i <= S; i += 2) if (!sieve[i]) {
